Move resize handling out of entry
