+++ src/index.ts
@@ -6,47 +6,10 @@
  * See the file LICENSE.txt for more information.
  */
 
-import Blits from "@lightningjs/blits";
-
-import LightningApp from "./LightningApp";
-import { debounce } from "./utils/debounce";
+import { initializeApp } from "./launcher";
 
 /**
- * Launch the LightningJS application sized to the current viewport.
+ * Application entry point. Launches the LightningJS view and hooks window
+ * resize events so the app always fills the viewport.
  */
-function launchLightningApp(width: number, height: number): void {
-  Blits.Launch(LightningApp, "app", {
-    w: width,
-    h: height,
-  });
-}
-
-/** Milliseconds to wait before applying the final size after a resize */
-const COOL_DOWN_MS: number = 100;
-
-/** Launch the app, replacing any existing canvas */
-function startApp(width: number, height: number): void {
-  const mount: HTMLElement = document.getElementById("app") as HTMLElement;
-  const oldCanvas: HTMLCanvasElement | null = mount.querySelector("canvas");
-
-  // Launch the new LightningJS canvas before removing the old one to minimize
-  // the time the screen goes blank during a resize
-  launchLightningApp(width, height);
-
-  if (oldCanvas !== null) {
-    oldCanvas.remove();
-  }
-}
-
-const debouncedStartApp: (...errArgs: Parameters<typeof startApp>) => void =
-  debounce(startApp, COOL_DOWN_MS);
-
-window.addEventListener("resize", (): void => {
-  debouncedStartApp(window.innerWidth, window.innerHeight);
-});
-
-/**
- * Application entry point. Loads global state and launches the LightningJS
- * view.
- */
-startApp(window.innerWidth, window.innerHeight);
+initializeApp();
